Norsk Creation Culture
This is a reference to FionaDanger's map of the area, which had the "Norsk" as a seperate culture in the region. I also figure this will be pretty frequent, so it'd be good to give it a name for when it does happen so it doesn't get a stupid name.